+++ airflow/example_dags/example_complex.py
@@ -23,9 +23,9 @@ from __future__ import annotations
 
 import pendulum
 
-from airflow.models.baseoperator import chain
 from airflow.models.dag import DAG
 from airflow.providers.standard.operators.bash import BashOperator
+from airflow.sdk import chain
 
 with DAG(
     dag_id="example_complex",
