+++ airflow/utils/edgemodifier.py
@@ -16,17 +16,5 @@
 # under the License.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
-import airflow.sdk
-
-if TYPE_CHECKING:
-    from airflow.typing_compat import TypeAlias
-
-EdgeModifier: TypeAlias = airflow.sdk.definitions.edges.EdgeModifier
-
-
-# Factory functions
-def Label(label: str):
-    """Create an EdgeModifier that sets a human-readable label on the edge."""
-    return EdgeModifier(label=label)
+# Re-export for compat
+from airflow.sdk.definitions.edges import Label as Label
