+++ newsfragments/aip-72.significant.rst
@@ -34,7 +34,7 @@ As part of this change the following breaking changes have occurred:
 
   .. code-block:: bash
 
-      airflow db drop-archived -t "_xcAm_archive"
+      airflow db drop-archived -t "_xcom_archive"
 
 - The ability to specify scheduling conditions for an operator via the ``deps`` class attribute has been removed.
 
@@ -46,7 +46,37 @@ As part of this change the following breaking changes have occurred:
 
   Any occurrences of imports from ``airflow.models.baseoperatorlink`` will need to be updated to ``airflow.sdk.definitions.baseoperatorlink``
 
-- With the We have removed DAG level settings that control the UI behaviour.
+- ``chain``, ``chain_linear`` and ``cross_downstream`` have been moved to the task SDK.
+
+  Any occurrences of imports from ``airflow.models.baseoperator`` will need to be updated to ``airflow.sdk``
+
+  Old imports:
+
+  .. code-block:: python
+
+      from airflow.models.baseoperator import chain, chain_linear, cross_downstream
+
+  New imports:
+
+  .. code-block:: python
+
+      from airflow.sdk import chain, chain_linear, cross_downstream
+
+- The ``Label`` class has been moved to the task SDK.
+
+  Old imports:
+
+  .. code-block:: python
+
+      from airflow.utils.edgemodifier import Label
+
+  New imports:
+
+  .. code-block:: python
+
+      from airflow.sdk import Label
+
+- We have removed DAG level settings that control the UI behaviour.
   These are now as per-user settings controlled by the UI
 
   - ``default_view``
