+++ tests/models/test_baseoperator.py
@@ -25,12 +25,9 @@ from unittest import mock
 import pytest
 
 from airflow.decorators import task as task_decorator
-from airflow.exceptions import AirflowException, TaskDeferralTimeout
+from airflow.exceptions import TaskDeferralTimeout
 from airflow.models.baseoperator import (
     BaseOperator,
-    chain,
-    chain_linear,
-    cross_downstream,
 )
 from airflow.models.dag import DAG
 from airflow.models.dagrun import DagRun
@@ -38,7 +35,6 @@ from airflow.models.taskinstance import TaskInstance
 from airflow.models.trigger import TriggerFailureReason
 from airflow.providers.common.compat.lineage.entities import File
 from airflow.providers.common.sql.operators import sql
-from airflow.utils.edgemodifier import Label
 from airflow.utils.task_group import TaskGroup
 from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.types import DagRunType
@@ -91,89 +87,6 @@ class TestBaseOperator:
             task_id="test_valid_trigger_rule", dag=non_fail_fast_dag, trigger_rule=TriggerRule.ALWAYS
         )
 
-    def test_cross_downstream(self):
-        """Test if all dependencies between tasks are all set correctly."""
-        dag = DAG(dag_id="test_dag", schedule=None, start_date=datetime.now())
-        start_tasks = [BaseOperator(task_id=f"t{i}", dag=dag) for i in range(1, 4)]
-        end_tasks = [BaseOperator(task_id=f"t{i}", dag=dag) for i in range(4, 7)]
-        cross_downstream(from_tasks=start_tasks, to_tasks=end_tasks)
-
-        for start_task in start_tasks:
-            assert set(start_task.get_direct_relatives(upstream=False)) == set(end_tasks)
-
-        # Begin test for `XComArgs`
-        xstart_tasks = [
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(1, 4)
-        ]
-        xend_tasks = [
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(4, 7)
-        ]
-        cross_downstream(from_tasks=xstart_tasks, to_tasks=xend_tasks)
-
-        for xstart_task in xstart_tasks:
-            assert set(xstart_task.operator.get_direct_relatives(upstream=False)) == {
-                xend_task.operator for xend_task in xend_tasks
-            }
-
-    def test_chain(self):
-        dag = DAG(dag_id="test_chain", schedule=None, start_date=datetime.now())
-
-        # Begin test for classic operators with `EdgeModifiers`
-        [label1, label2] = [Label(label=f"label{i}") for i in range(1, 3)]
-        [op1, op2, op3, op4, op5, op6] = [BaseOperator(task_id=f"t{i}", dag=dag) for i in range(1, 7)]
-        chain(op1, [label1, label2], [op2, op3], [op4, op5], op6)
-
-        assert {op2, op3} == set(op1.get_direct_relatives(upstream=False))
-        assert [op4] == op2.get_direct_relatives(upstream=False)
-        assert [op5] == op3.get_direct_relatives(upstream=False)
-        assert {op4, op5} == set(op6.get_direct_relatives(upstream=True))
-
-        assert dag.get_edge_info(upstream_task_id=op1.task_id, downstream_task_id=op2.task_id) == {
-            "label": "label1"
-        }
-        assert dag.get_edge_info(upstream_task_id=op1.task_id, downstream_task_id=op3.task_id) == {
-            "label": "label2"
-        }
-
-        # Begin test for `XComArgs` with `EdgeModifiers`
-        [xlabel1, xlabel2] = [Label(label=f"xcomarg_label{i}") for i in range(1, 3)]
-        [xop1, xop2, xop3, xop4, xop5, xop6] = [
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(1, 7)
-        ]
-        chain(xop1, [xlabel1, xlabel2], [xop2, xop3], [xop4, xop5], xop6)
-
-        assert {xop2.operator, xop3.operator} == set(xop1.operator.get_direct_relatives(upstream=False))
-        assert [xop4.operator] == xop2.operator.get_direct_relatives(upstream=False)
-        assert [xop5.operator] == xop3.operator.get_direct_relatives(upstream=False)
-        assert {xop4.operator, xop5.operator} == set(xop6.operator.get_direct_relatives(upstream=True))
-
-        assert dag.get_edge_info(
-            upstream_task_id=xop1.operator.task_id, downstream_task_id=xop2.operator.task_id
-        ) == {"label": "xcomarg_label1"}
-        assert dag.get_edge_info(
-            upstream_task_id=xop1.operator.task_id, downstream_task_id=xop3.operator.task_id
-        ) == {"label": "xcomarg_label2"}
-
-        # Begin test for `TaskGroups`
-        [tg1, tg2] = [TaskGroup(group_id=f"tg{i}", dag=dag) for i in range(1, 3)]
-        [op1, op2] = [BaseOperator(task_id=f"task{i}", dag=dag) for i in range(1, 3)]
-        [tgop1, tgop2] = [
-            BaseOperator(task_id=f"task_group_task{i}", task_group=tg1, dag=dag) for i in range(1, 3)
-        ]
-        [tgop3, tgop4] = [
-            BaseOperator(task_id=f"task_group_task{i}", task_group=tg2, dag=dag) for i in range(1, 3)
-        ]
-        chain(op1, tg1, tg2, op2)
-
-        assert {tgop1, tgop2} == set(op1.get_direct_relatives(upstream=False))
-        assert {tgop3, tgop4} == set(tgop1.get_direct_relatives(upstream=False))
-        assert {tgop3, tgop4} == set(tgop2.get_direct_relatives(upstream=False))
-        assert [op2] == tgop3.get_direct_relatives(upstream=False)
-        assert [op2] == tgop4.get_direct_relatives(upstream=False)
-
     def test_baseoperator_raises_exception_when_task_id_plus_taskgroup_id_exceeds_250_chars(self):
         """Test exception is raised when operator task id + taskgroup id > 250 chars."""
         dag = DAG(dag_id="foo", schedule=None, start_date=datetime.now())
@@ -200,110 +113,6 @@ class TestBaseOperator:
             BaseOperator(task_id="1" * 249, dag=dag)
         except Exception as e:
             pytest.fail(f"Exception raised: {e}")
-
-    def test_chain_linear(self):
-        dag = DAG(dag_id="test_chain_linear", schedule=None, start_date=datetime.now())
-
-        t1, t2, t3, t4, t5, t6, t7 = (BaseOperator(task_id=f"t{i}", dag=dag) for i in range(1, 8))
-        chain_linear(t1, [t2, t3, t4], [t5, t6], t7)
-
-        assert set(t1.get_direct_relatives(upstream=False)) == {t2, t3, t4}
-        assert set(t2.get_direct_relatives(upstream=False)) == {t5, t6}
-        assert set(t3.get_direct_relatives(upstream=False)) == {t5, t6}
-        assert set(t7.get_direct_relatives(upstream=True)) == {t5, t6}
-
-        t1, t2, t3, t4, t5, t6 = (
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(1, 7)
-        )
-        chain_linear(t1, [t2, t3], [t4, t5], t6)
-
-        assert set(t1.operator.get_direct_relatives(upstream=False)) == {t2.operator, t3.operator}
-        assert set(t2.operator.get_direct_relatives(upstream=False)) == {t4.operator, t5.operator}
-        assert set(t3.operator.get_direct_relatives(upstream=False)) == {t4.operator, t5.operator}
-        assert set(t6.operator.get_direct_relatives(upstream=True)) == {t4.operator, t5.operator}
-
-        # Begin test for `TaskGroups`
-        tg1, tg2 = (TaskGroup(group_id=f"tg{i}", dag=dag) for i in range(1, 3))
-        op1, op2 = (BaseOperator(task_id=f"task{i}", dag=dag) for i in range(1, 3))
-        tgop1, tgop2 = (
-            BaseOperator(task_id=f"task_group_task{i}", task_group=tg1, dag=dag) for i in range(1, 3)
-        )
-        tgop3, tgop4 = (
-            BaseOperator(task_id=f"task_group_task{i}", task_group=tg2, dag=dag) for i in range(1, 3)
-        )
-        chain_linear(op1, tg1, tg2, op2)
-
-        assert set(op1.get_direct_relatives(upstream=False)) == {tgop1, tgop2}
-        assert set(tgop1.get_direct_relatives(upstream=False)) == {tgop3, tgop4}
-        assert set(tgop2.get_direct_relatives(upstream=False)) == {tgop3, tgop4}
-        assert set(tgop3.get_direct_relatives(upstream=False)) == {op2}
-        assert set(tgop4.get_direct_relatives(upstream=False)) == {op2}
-
-        t1, t2 = (BaseOperator(task_id=f"t-{i}", dag=dag) for i in range(1, 3))
-        with pytest.raises(ValueError, match="Labels are not supported"):
-            chain_linear(t1, Label("hi"), t2)
-
-        with pytest.raises(ValueError, match="nothing to do"):
-            chain_linear()
-
-        with pytest.raises(ValueError, match="Did you forget to expand"):
-            chain_linear(t1)
-
-    def test_chain_not_support_type(self):
-        dag = DAG(dag_id="test_chain", schedule=None, start_date=datetime.now())
-        [op1, op2] = [BaseOperator(task_id=f"t{i}", dag=dag) for i in range(1, 3)]
-        with pytest.raises(TypeError):
-            chain([op1, op2], 1)
-
-        # Begin test for `XComArgs`
-        [xop1, xop2] = [
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(1, 3)
-        ]
-
-        with pytest.raises(TypeError):
-            chain([xop1, xop2], 1)
-
-        # Begin test for `EdgeModifiers`
-        with pytest.raises(TypeError):
-            chain([Label("labe1"), Label("label2")], 1)
-
-        # Begin test for `TaskGroups`
-        [tg1, tg2] = [TaskGroup(group_id=f"tg{i}", dag=dag) for i in range(1, 3)]
-
-        with pytest.raises(TypeError):
-            chain([tg1, tg2], 1)
-
-    def test_chain_different_length_iterable(self):
-        dag = DAG(dag_id="test_chain", schedule=None, start_date=datetime.now())
-        [label1, label2] = [Label(label=f"label{i}") for i in range(1, 3)]
-        [op1, op2, op3, op4, op5] = [BaseOperator(task_id=f"t{i}", dag=dag) for i in range(1, 6)]
-
-        with pytest.raises(AirflowException):
-            chain([op1, op2], [op3, op4, op5])
-
-        with pytest.raises(AirflowException):
-            chain([op1, op2, op3], [label1, label2])
-
-        # Begin test for `XComArgs` with `EdgeModifiers`
-        [label3, label4] = [Label(label=f"xcomarg_label{i}") for i in range(1, 3)]
-        [xop1, xop2, xop3, xop4, xop5] = [
-            task_decorator(task_id=f"xcomarg_task{i}", python_callable=lambda: None, dag=dag)()
-            for i in range(1, 6)
-        ]
-
-        with pytest.raises(AirflowException):
-            chain([xop1, xop2], [xop3, xop4, xop5])
-
-        with pytest.raises(AirflowException):
-            chain([xop1, xop2, xop3], [label1, label2])
-
-        # Begin test for `TaskGroups`
-        [tg1, tg2, tg3, tg4, tg5] = [TaskGroup(group_id=f"tg{i}", dag=dag) for i in range(1, 6)]
-
-        with pytest.raises(AirflowException):
-            chain([tg1, tg2], [tg3, tg4, tg5])
 
     def test_lineage_composition(self):
         """
