+++ tests/system/example_empty.py
@@ -18,9 +18,9 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from airflow.models.baseoperator import chain
 from airflow.models.dag import DAG
 from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.sdk import chain
 
 DAG_ID = "example_empty"
